vehicles/requirements.txt
grpcio==1.68.1
grpcio-tools==1.68.1
grpcio-reflection==1.68.1
SQLAlchemy==2.0.36
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn[standard]==0.32.1
aiofiles==24.1.0
pytest==8.3.4
pytest-mock==3.14.0
